requirements.txt
web3==5.25.0
aiohttp==3.8.1
git+https://github.com/Pycord-Development/pycord.git
aiodns==3.0.0
python-dotenv==0.19.2
motor==2.5.1
dnspython==2.1.0
certifi==2021.10.8
uvloop==0.16.0; sys_platform == 'linux'
pytimeparse==1.1.8